+++ examples/analysis.py
@@ -10,7 +10,10 @@ import matplotlib.pyplot as plt
 import matplotlib.animation as animation
 import mpl_toolkits.mplot3d as mp3d
 import numpy as np
+import pandas as pd
 import seaborn as sns
+import sklearn.metrics
+import sklearn.neighbors
 import torch
 import wandb
 
@@ -24,8 +27,8 @@ import inept
 # Get args
 import sys
 # run_id_idx = int(sys.argv[1])
-# analysis_key_idx = int(sys.argv[1])
-stage_override = int(sys.argv[1])
+analysis_key_idx = int(sys.argv[1])
+# stage_override = int(sys.argv[1])
 
 # Set params
 DEVICE = 'cuda:0' if torch.cuda.is_available() else 'cpu'
@@ -59,7 +62,7 @@ torch.set_grad_enabled(False);
 
 # %%
 # Parameters
-run_id_idx = 1
+run_id_idx = 3
 run_id = (
     'brf6n6sn',  # TemporalBrain Random 100 Max
     'rypltvk5',  # MMD-MA Random 100 Max (requires `total_statistics`)
@@ -70,13 +73,14 @@ run_id = (
     'vb1x7bae',  # MERFISH NR
     '473vyon2',  # ISS NR
 )[run_id_idx]
-# stage_override = None  # Manually override policy stage selection
+stage_override = None  # Manually override policy stage selection
 num_nodes_override = None
 max_batch_override = None
 max_nodes_override = None
 seed_override = None
 
 # Load run
+print(f'Loading run {run_id}...')
 api = wandb.Api()
 run = api.run(f'oafish/INEPT/{run_id}')
 config = defaultdict(lambda: {})
@@ -93,13 +97,15 @@ if torch.cuda.is_available(): torch.cuda.manual_seed(seed)
 np.random.seed(seed)
 
 # Load data
+print(f'Loading dataset {config["data"]["dataset"]}...')
 modalities, types, features = data.load_data(config['data']['dataset'], DATA_FOLDER)
 # config['data'] = inept.utilities.overwrite_dict(config['data'], {'standardize': True})  # Old model compatibility
 # config['data'] = inept.utilities.overwrite_dict(config['data'], {'top_variant': [20, 20]})  # Top variant testing
 if num_nodes_override is not None: config['data'] = inept.utilities.overwrite_dict(config['data'], {'num_nodes': num_nodes_override})
 if max_batch_override is not None: config['train'] = inept.utilities.overwrite_dict(config['train'], {'max_batch': max_batch_override})
 ppc = inept.utilities.Preprocessing(**config['data'], device=DEVICE)
-modalities, features = ppc.fit_transform(modalities, features, total_statistics=True)
+total_statistics = False  # Legacy compatibility
+modalities, features = ppc.fit_transform(modalities, features, total_statistics=total_statistics)
 modalities, types = ppc.subsample(modalities, types)
 modalities = ppc.cast(modalities)
 
@@ -151,51 +157,13 @@ labels = types[0][:, 0]
 times = types[0][:, -1]  # Temporary time annotation, will change per-dataset
 
 # %% [markdown]
-# ### Static Visualizations
-
-# %%
-# Load history from wandb
-history = run.history(samples=2000)
-history['timestep'] = history['end_timestep']
-history['Runtime (h)'] = history['_runtime'] / 60**2
-
-# Plot
-fig, ax = plt.subplots(1, 1, figsize=(8, 3), sharex=True, layout='constrained')
-def plot_without_zeros(x, y, **kwargs):
-    x, y = x[np.argwhere(y != 0).flatten()], y[np.argwhere(y != 0).flatten()]
-    ax.plot(x, y, **kwargs)
-ax.plot(history['timestep'], history['average_reward'], color='black', lw=3, label='Average Reward')
-plot_without_zeros(history['timestep'], history['rewards/bound'], color='red', alpha=.75, lw=2, label='Boundary Penalty')
-plot_without_zeros(history['timestep'], history['rewards/velocity'], color='goldenrod', alpha=.75, lw=2, label='Velocity Penalty')
-plot_without_zeros(history['timestep'], history['rewards/action'], color='green', alpha=.75, lw=2, label='Action Penalty')
-plot_without_zeros(history['timestep'], history['rewards/distance'], color='blue', alpha=.75, lw=2, label='Distance Reward')
-plot_without_zeros(history['timestep'], history['rewards/origin'], color='darkorange', alpha=.75, lw=2, label='Origin Reward')
-
-# Stage ticks
-unique, stage_idx = np.unique(history['stage'][::-1], return_index=True)
-stage_idx = len(history['stage']) - stage_idx
-stage_idx = stage_idx[:-1]
-[ax.axvline(x=history['timestep'][idx], color='black', alpha=.5, linestyle='dashed', lw=1) for idx in stage_idx]
-
-# Labels
-ax.set_xlabel('Timestep')
-ax.set_ylabel('Reward')
-ax.legend(ncols=3)
-
-# Styling
-ax.spines[['right', 'top']].set_visible(False)
-
-# Save plot
-fname = f'{config["data"]["dataset"]}_performance.pdf'
-fig.savefig(os.path.join(PLOT_FOLDER, fname), dpi=300)
-
-# %% [markdown]
 # ### Generate Runs
 
 # %%
 # Choose key
 # TODO: Calculate all, plot one (?)
-analysis_key_idx = 0
+# analysis_key_idx = 0
+optimize_memory = True  # Saves memory by shrinking env based on present
 discovery_key = 0  # Auto
 temporal_key = 0  # Auto
 perturbation_features = [np.random.choice(len(fs), 10, replace=False) for i, fs in enumerate(features) if (i not in env.reward_distance_target) or (len(env.reward_distance_target) == len(modalities))]
@@ -223,7 +191,7 @@ discovery = discovery[discovery_key]
 # Stage order list
 temporal = []
 # Reverse alphabetical (ExSeq, MERFISH, smFISH, ISS, MouseVisual)
-temporal_general = {'stages': [[l] for l in np.unique(labels)[::-1]]}
+temporal_general = {'stages': [[l] for l in np.unique(times)[::-1]]}
 temporal += [temporal_general]
 # Choose stage order
 temporal = temporal[temporal_key]
@@ -235,6 +203,10 @@ perturbation_feature_names = [[fnames[pf] for pf in pfs] for pfs, fnames in zip(
 memories = {}
 
 # %%
+# Profiling
+profile = True
+if profile: torch.cuda.memory._record_memory_history(max_entries=100000)
+
 # Choose state manager
 state_manager = state_manager_class(
     device=DEVICE,
@@ -245,27 +217,31 @@ state_manager = state_manager_class(
     num_nodes=modalities[0].shape[0],
     dim=env.dim,
 )
+# TODO: Figure out why -1 on repeat for temporal/perturbation?
 get_current_stage = lambda: state_manager.current_stage if state_manager_class in (inept.utilities.TemporalStateManager, inept.utilities.PerturbationStateManager) else -1
 
 # Initialize
-env.reset(); memories[analysis_key] = defaultdict(lambda: [])
+env.set_modalities(modalities); env.reset(); memories[analysis_key] = defaultdict(lambda: [])
 
 # Modify
 state_vars, end = state_manager(
     # present=present,
     state=env.get_state(),
-    modalities=ppc.cast(ppc.inverse_transform(ppc.inverse_cast(env.get_modalities()))),
+    # TODO: Somehow replace the inverse_transform with just the original data? For sparse compatibility
+    modalities=ppc.cast(ppc.inverse_transform(ppc.inverse_cast(modalities)), device='cpu'),
     labels=labels,
     times=times,
 )
 present = state_vars['present']
-env.set_state(state_vars['state'])
+memory_mask = present if optimize_memory else torch.ones_like(present, device=DEVICE)
+full_state = state_vars['state']
+env.set_state(full_state[memory_mask])
 raw_modalities = state_vars['modalities']
-env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast(raw_modalities))))
+env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast([m[memory_mask.cpu()] for m in raw_modalities]))))
 
 # Continue initializing
 memories[analysis_key]['present'].append(present)
-memories[analysis_key]['states'].append(env.get_state())
+memories[analysis_key]['states'].append(full_state.clone())
 memories[analysis_key]['stages'].append(get_current_stage())
 memories[analysis_key]['rewards'].append(torch.zeros(modalities[0].shape[0], device=DEVICE))
 
@@ -282,35 +258,41 @@ while True:
     state = env.get_state(include_modalities=True)
     actions = torch.zeros((modalities[0].shape[0], env.dim), device=DEVICE)
     actions[present] = policy.act_macro(
-        state[present],
+        state if optimize_memory else state[present],
         keys=torch.arange(modalities[0].shape[0], device=DEVICE)[present],
         max_batch=config['train']['max_batch'],
     )
-    rewards = torch.zeros(modalities[0].shape[0])
-    # TODO: Currently, rewards factor in non-present nodes
-    rewards, _, _ = env.step(actions, return_itemized_rewards=True)
-    new_state = env.get_state()
-    new_state[~present] = state[~present, :2*env.dim]  # Don't move un-spawned nodes
-    env.set_state(new_state)
+    rewards = torch.zeros(modalities[0].shape[0], device=DEVICE)
+    new_rewards, _, _ = env.step(actions[present] if optimize_memory else actions, return_itemized_rewards=True)
+    if optimize_memory: rewards[present] = new_rewards
+    else: rewards = new_rewards
+    full_state[present] = env.get_state() if optimize_memory else env.get_state()[present]
+    if not optimize_memory: env.set_state(full_state)  # Don't move un-spawned nodes
 
     # Modify
     state_vars, end = state_manager(
         present=present,
-        state=env.get_state(),
+        state=full_state,
         modalities=raw_modalities,
         labels=labels,
         times=times,
     )
+    present_change = (state_vars['present'] != present).any()
     present = state_vars['present']
-    env.set_state(state_vars['state'])
+    memory_mask = present if optimize_memory else torch.ones_like(present, device=DEVICE)
+    full_state = state_vars['state']
+    env.set_state(full_state[memory_mask])
     # Only modify if changes
-    if torch.tensor([(rm != svm).any() for rm, svm in zip(raw_modalities, state_vars['modalities'])]).any():
+    if (
+        torch.tensor([(rm != svm).any() for rm, svm in zip(raw_modalities, state_vars['modalities'])]).any()
+        or (optimize_memory and present_change)
+    ):
         raw_modalities = state_vars['modalities']
-        env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast(raw_modalities))))
+        env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast([m[memory_mask.cpu()] for m in raw_modalities]))))
 
     # Record
     memories[analysis_key]['present'].append(present)
-    memories[analysis_key]['states'].append(env.get_state())
+    memories[analysis_key]['states'].append(full_state.clone())
     memories[analysis_key]['stages'].append(get_current_stage())
     memories[analysis_key]['rewards'].append(rewards)
 
@@ -325,17 +307,22 @@ memories[analysis_key]['stages'] = torch.tensor(memories[analysis_key]['stages']
 memories[analysis_key]['rewards'] = torch.stack(memories[analysis_key]['rewards'])
 memories[analysis_key] = dict(memories[analysis_key])
 
+# Profiling
+if profile:
+    torch.cuda.memory._dump_snapshot('snapshot.pickle')
+    torch.cuda.memory._record_memory_history(enabled=None)
+
 # CLI
 print()
+
+# %%
+memories[analysis_key]['rewards'].cpu().mean()
 
 # %%
 stages, counts = np.unique(memories[analysis_key]['stages'], return_counts=True)
 print('Steps per Stage:')
 for s, c in zip(stages, counts):
     print(f'{s}: {c}')
-
-# %%
-memories[analysis_key]['rewards'].cpu().mean()  # TODO: Why not the same as WANDB?
 
 # %%
 # Perturbation significance analysis
@@ -377,7 +364,147 @@ if 'perturbation' in memories:
         print()
 
 # %% [markdown]
-# ### Plot Memories
+# ### Static Visualizations
+
+# %% [markdown]
+# ##### Loss Plot
+
+# %%
+# Load history from wandb
+history = run.history(samples=2000)
+history['timestep'] = history['end_timestep']
+history['Runtime (h)'] = history['_runtime'] / 60**2
+
+# Plot
+fig, ax = plt.subplots(1, 1, figsize=(18, 6), layout='constrained')
+def plot_without_zeros(x, y, **kwargs):
+    x, y = x[np.argwhere(y != 0).flatten()], y[np.argwhere(y != 0).flatten()]
+    ax.plot(x, y, **kwargs)
+ax.plot(history['timestep'], history['average_reward'], color='black', lw=3, label='Average Reward')
+plot_without_zeros(history['timestep'], history['rewards/bound'], color='red', alpha=.75, lw=2, label='Boundary Penalty')
+plot_without_zeros(history['timestep'], history['rewards/velocity'], color='goldenrod', alpha=.75, lw=2, label='Velocity Penalty')
+plot_without_zeros(history['timestep'], history['rewards/action'], color='green', alpha=.75, lw=2, label='Action Penalty')
+plot_without_zeros(history['timestep'], history['rewards/distance'], color='blue', alpha=.75, lw=2, label='Distance Reward')
+plot_without_zeros(history['timestep'], history['rewards/origin'], color='darkorange', alpha=.75, lw=2, label='Origin Reward')
+
+# Stage ticks
+unique, stage_idx = np.unique(history['stage'][::-1], return_index=True)
+stage_idx = len(history['stage']) - stage_idx
+stage_idx = stage_idx[:-1]
+[ax.axvline(x=history['timestep'][idx], color='black', alpha=.5, linestyle='dashed', lw=1) for idx in stage_idx]
+
+# Labels
+ax.set_xlabel('Timestep')
+ax.set_ylabel('Reward')
+ax.legend(loc='lower right', ncols=3)
+
+# Styling
+ax.spines[['right', 'top']].set_visible(False)
+ax.set_xlim([0, history['timestep'].max()])
+
+# Save plot
+fname = f'{config["data"]["dataset"]}_performance.pdf'
+fig.savefig(os.path.join(PLOT_FOLDER, fname), dpi=300)
+
+# %% [markdown]
+# ##### Comparison
+
+# %%
+# Method comparison
+if 'integration' in memories:
+    # Comparison metrics
+    metric_rand = lambda X: np.random.rand()
+    metric_silhouette = lambda X: sklearn.metrics.silhouette_score(X, labels)
+    metric_ch_score = lambda X: sklearn.metrics.calinski_harabasz_score(X, labels)
+    def metric_knn_ami(X):
+        knn = sklearn.neighbors.KNeighborsClassifier(n_neighbors=10)
+        knn.fit(X, labels)
+        pred = knn.predict(X)
+        return sklearn.metrics.adjusted_mutual_info_score(labels, pred)
+
+    metric_tuples = {
+        'rand': (metric_rand, {'label': 'Random'}),
+        'sc': (metric_silhouette, {'label': 'Silhouette Coefficient'}),
+        'knn_ami': (metric_knn_ami, {'label': 'KNN Adjusted Mutual Information'}),
+        'ch': (metric_ch_score, {'label': 'Calinski Harabasz Index', 'scale': 'log'}),
+    }
+
+    # Select metrics
+    metric_x, kwargs_x = metric_tuples['ch']
+    metric_y, kwargs_y = metric_tuples['knn_ami']
+
+    # Get other methods
+    method_dir = os.path.join(BASE_FOLDER, '../other_methods/runs', config['data']['dataset'])
+    method_names = next(os.walk(method_dir))[1]
+    method_results = {}
+    for name in method_names:
+        # Get output files
+        files = os.listdir(os.path.join(method_dir, name))
+        r = re.compile('^P\d+.txt$')
+        files = list(filter(r.match, files))
+
+        # Record
+        for i, file in enumerate(files):
+            proj = np.loadtxt(os.path.join(method_dir, name, file))
+            method_results[(name, i)] = proj
+
+    # Add cellTRIP
+    method_results[('cellTRIP', -1)] = memories['integration']['states'][-1].detach().cpu()
+
+    # Compile and calculate performances
+    performance = pd.DataFrame(columns=['Method', 'Modality', 'x', 'y'])
+    for key, data in method_results.items():
+        performance.loc[performance.shape[0]] = [key[0], key[1], metric_x(data), metric_y(data)]
+
+    # Plot with text
+    fig, ax = plt.subplots(1, 1, figsize=(6, 6), sharex=True, layout='constrained')
+    method_colors = {}
+    annotations = []
+    for i, r in performance.iterrows():
+        # Set color
+        if r['Method'] not in method_colors: method_colors[r['Method']] = sns.color_palette()[len(method_colors)]
+        
+        # Plot
+        ax.scatter(
+            r['x'],
+            r['y'],
+            color=method_colors[r['Method']],
+            s=100,
+        )
+
+        # Cross lines
+        ax.axvline(x=r['x'], ls='--', alpha=.1, color='black', zorder=.3)
+        ax.axhline(y=r['y'], ls='--', alpha=.1, color='black', zorder=.3)
+
+        # Annotate
+        text = f'{r["Method"]}' + (f' ({r["Modality"]})' if r['Modality'] != -1 else '')
+        annotations.append(ax.text(
+            r['x'], r['y'], text,
+            ha='center', va='center', fontsize='large'))
+
+    # Styling
+    ax.spines[['right', 'top', 'bottom', 'left']].set_visible(False)
+    ax.set(
+        **{'x'+k: v for k, v in kwargs_x.items()},
+        **{'y'+k: v for k, v in kwargs_y.items()},
+    )
+    ax.axvline(x=0, ls='-', alpha=.6, color='black', zorder=.1)
+    ax.axhline(y=0, ls='-', alpha=.6, color='black', zorder=.1)
+
+    # Adjust Annotation Positions
+    from adjustText import adjust_text
+    adjust_text(
+        annotations,
+        expand=(1.2, 2),
+        arrowprops=dict(arrowstyle='->', color='black', zorder=.3),
+    )
+
+    # Save plot
+    fname = f'{config["data"]["dataset"]}_comparison.pdf'
+    fig.savefig(os.path.join(PLOT_FOLDER, fname), dpi=300)
+
+# %% [markdown]
+# ### Dynamic Visualizations
 
 # %%
 # Prepare data
@@ -446,6 +573,8 @@ elif analysis_key in ('perturbation',):
     ]
 
 # Initialize views
+num_lines = 100
+if analysis_key == 'temporal': num_lines *= len(temporal['stages'])
 arguments = {
     # Data
     'present': present,
@@ -465,6 +594,7 @@ arguments = {
     'skip': skip,
     'seed': 42,
     # Styling
+    'num_lines': num_lines,
     'ms': 5,  # 3
     'lw': 1,
 }